Refactor to better support usage as module (#61)
* Refactor to better support usage as module

* Bump version

* Ensure only 1 connection during lambda execution

* Fix typing

* Rework to lazy-app to make for easier testing

* Update path

* Cleanup helper for flexibility

* Use context manager to close connections

* Add tests

* Pre-commit

* Rework settings

* Simplify db setup

* Add missing quote and change steps to load data

* Fix settings import

* Update for fast reloading and how to run API locally

* Update README.md


---------
