+++ space2stats_api/src/space2stats/__main__.py
@@ -1,7 +1,5 @@
 import os
 
-from .app import app
-
 try:
     import uvicorn  # noqa
 
@@ -15,9 +13,11 @@ if __name__ == "__main__":
     ), "uvicorn must be installed: `python -m pip install 'space2stats[server]'`"
 
     uvicorn.run(
-        app=app,
+        app="space2stats.api.app:build_app",
         host=os.getenv("UVICORN_HOST", "127.0.0.1"),
-        port=os.getenv("UVICORN_PORT", "8000"),
+        port=int(os.getenv("UVICORN_PORT", "8000")),
         root_path=os.getenv("UVICORN_ROOT_PATH", ""),
         log_level="info",
+        factory=True,
+        reload=True,
     )
