+++ space2stats_api/src/space2stats/api/__init__.py
@@ -1,0 +1,3 @@
+from .app import build_app
+
+__all__ = ["build_app"]
