+++ space2stats_api/src/space2stats/api/db.py
@@ -10,13 +10,10 @@ from .settings import Settings
 
 async def connect_to_db(
     app: FastAPI,
-    settings: Optional[Settings] = None,
+    settings: Settings,
     pool_kwargs: Optional[Dict[str, Any]] = None,
 ) -> None:
     """Connect to Database."""
-    if not settings:
-        settings = Settings()
-
     pool_kwargs = pool_kwargs or {}
 
     app.state.pool = ConnectionPool(
