+++ space2stats_api/src/space2stats/api/settings.py
@@ -1,0 +1,6 @@
+from ..settings import Settings as DbSettings
+
+
+class Settings(DbSettings):
+    # Bucket for large responses
+    S3_BUCKET_NAME: str
