+++ space2stats_api/src/space2stats/types.py
@@ -1,0 +1,6 @@
+from typing import Dict
+
+from geojson_pydantic import Feature, Polygon
+from typing_extensions import TypeAlias
+
+AoiModel: TypeAlias = Feature[Polygon, Dict]
